rag-system/backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-mistralai
faiss-cpu
python-dotenv
python-multipart
pypdf
unstructured
unstructured-inference
